Ajout d'un volume a chaque carton (correction)

+++ class/Carton.h
@@ -13,13 +13,13 @@ public:
     int getLargeur() { return _largeur; }
     int getLongueur() { return _longueur; }
     int getHauteur() { return _hauteur; }
-    int getVolume() { return _volume; }
+    double getVolume() { return _volume; }
 
 private:
     int _largeur;
     int _longueur;
     int _hauteur;
-    int _volume;
+    double _volume;
 };
 
 #endif //PROJET_CARTON_H
